docs: fix build warning (collected multiple info for make-mapping-change-listener)
Both the defstruct[*] forms and the defproc for mapping-change-listener
collect information on the name mapping-change-listener (because the
struct forms assume a constructor with that name). Combinations of
#:omit-constructor or #:constructor-name <mapping-change-listener> and
keeping or removing #:prefab didn't remove the error (while acting
slightly more honest), so instead document the 2 main procedures as if
they were part of a data definition (which they are). This is, after
all, the way they are provided from high-level.rkt.

This does remove the #:prefab documentation, but the struct was never
useful as a prefab anyway (you can't read/write thread handles, and
read/write are the main use of #:prefab). Perhaps #:transparent would be
a better marker? Using #:transparent #:omit-constructor in the docs also
doesn't prevent the issue, however.